requirements.txt
transformers==4.30.2
pillow==9.5.0
pandas==2.0.3
numpy==1.25.0
tqdm==4.65.0